Move sample_manager to seperate directory
Moved because it's a big part of the backend and not a small utils script anymore
