+++ AlgorithmAnalyzer/Backend/sample_manager/SampleManager.py
