+++ AlgorithmAnalyzer/Backend/sample_manager/tests.py
@@ -4,7 +4,7 @@ import shutil
 from pathlib import Path
 
 from werkzeug.datastructures import FileStorage
-from Backend.utils.sample_manager.SampleManager import SampleManager, UsernameException
+from Backend.sample_manager.SampleManager import SampleManager, UsernameException
 
 
 class TestSampleManager(unittest.TestCase):
